requirements.txt
Django==5.0.6
djangorestframework==3.15.2
transformers==4.44.2
torch>=2.1.0
librosa==0.10.2.post1
soundfile==0.12.1
numpy>=1.23
ffmpeg-python==0.2.0
uvicorn==0.30.6

